// engemedical-backend/src/soc/types/PedidoExame.ts
export type PedidoExameRequest = {
  empresa: string;
  funcionario?: string;
  dataInicio?: string;
  dataFim?: string;
  unidade?: string;
};

export type PedidoExame = {
  CODIGOEMPRESA: string;
  NOMEEMPRESA: string;
  CNPJEMPRESA: string;
  CPFEMPRESA: string;
  CEIEMPRESA: string;
  SUBGRUPOEMPRESA: string;
  SEQUENCIAFICHA: string;
  CODIGOFUNCIONARIO: string;
  NOMEFUNCIONARIO: string;
  CPFFUNCIONARIO: string;
  MATRICULAFUNCIONARIO: string;
  RGFUNCIONARIO: string;
  CODIGOCENTROCUSTO: string;
  NOMECENTROCUSTO?: string;
  CODIGOUNIDADE: string;
  CNPJUNIDADE: string;
  CPFUNIDADE: string;
  CEIUNIDADE: string;
  NOMEUNIDADE: string;
  CODIGOSETOR: string;
  NOMESETOR: string;
  CODIGOCARGO: string;
  NOMECARGO: string;
  DATACRIACAOPEDIDOEXAMES: string;
  CODIGOPRESTADOR: string;
  NOMEPRESTADOR: string;
  CODIGOTIPOEXAME: string;
  DATAFICHA: string;
  CODIGOTUSSEXAME: string;
  CODIGOEXAMEAMB: string;
  CODIGOINTERNOEXAME: string;
  NOMEEXAME: string;
  DATAEXAME: string;
  RISCOSFUNCIONARIO: string;
  MEDICOCOORDENADOR?: string;
  UFMEDICO?: string;
  CRM?: string;
  RESPONSAVELPEDIDOEXAME?: string;
  RISCOSASO: string;
  DATANASCIMENTO: string;
  SEXO?: string;
  NOMEDAMAEDOFUNCIONARIO: string;
  CPFMEDICOEXAMINADOR?: string;
  CODIGORH: string;
};